feat: 카테고리 칩을 추가합니다.
